pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "copom-rag-api"
version = "0.1.0"
description = "RAG API for querying COPOM documents stored in PostgreSQL+pgvector"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    "google-generativeai>=0.7.0",
    "psycopg2-binary>=2.9",
    "pgvector>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27", "ruff>=0.4"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
